+++ xharbour/source/rtl/cdpapi.c
@@ -1,5 +1,5 @@
 /*
- * $Id: cdpapi.c,v 1.28 2007/01/09 22:05:23 druzus Exp $
+ * $Id: cdpapi.c,v 1.29 2007/01/11 21:06:52 druzus Exp $
  */
 
 /*
@@ -576,7 +576,7 @@ HB_EXPORT ULONG hb_cdpUTF8ToStrn( PHB_CODEPAGE cdp, BOOL fCtrl,
                      }
                   }
                }
-               pDst[ ulD ] = uc >= 0x100 ? '?' : uc;
+               pDst[ ulD ] = uc >= 0x100 ? '?' : ( BYTE ) uc;
             }
             ++ulD;
          }
